Website now fetches posts, users and usergroups asynchronously from MySQL
The website is now integrated with MySQL, retrieving information
asynchronously using promises.
The User class and the Usergroup class have been added to serve as data
structures for information about the user and their usergroup
respectively. This is primarily to serve the name and color of a user,
however it is also serving the permissions a user holds aswell.